Add files via upload
Cooling shroud for Dragon hot end mounted on Jubilee tool plate.  Hot end needs to be rotated 90 degrees and the nozzle wiper will not allow the tool to fully dock.  Also, Y values in tfree and post will will need to be adjusted to keep the shroud from being pushed into the frame.  This will likely work for a rotated V6 nozzle as well.